+++ uikit-samples/src/main/java/com/sendbird/uikit/samples/BaseApplication.kt
@@ -65,6 +65,58 @@ class BaseApplication : MultiDexApplication() {
                     }
                 }
             }, context)
+
+            // set theme mode
+            SendbirdUIKit.setDefaultThemeMode(PreferenceUtils.themeMode)
+            // register push notification
+            SendbirdPushHelper.registerPushHandler(MyFirebaseMessagingService())
+            // set logger
+            SendbirdUIKit.setLogLevel(SendbirdUIKit.LogLevel.ALL)
+        }
+
+        /**
+         * In a sample app, different contextual settings are used in a single app.
+         * These are only used in the sample, because if the app kills and resurrects due to low memory, the last used sample settings should be preserved.
+         */
+        fun setupConfigurations() {
+            when (PreferenceUtils.selectedSampleType) {
+                SampleType.Basic -> {
+                    // set whether to use user profile
+                    UIKitConfig.common.enableUsingDefaultUserProfile = true
+                    // set whether to use typing indicators in channel list
+                    UIKitConfig.groupChannelListConfig.enableTypingIndicator = true
+                    // set whether to use read/delivery receipt in channel list
+                    UIKitConfig.groupChannelListConfig.enableMessageReceiptStatus = true
+                    // set whether to use user mention
+                    UIKitConfig.groupChannelConfig.enableMention = true
+                    // set reply type
+                    UIKitConfig.groupChannelConfig.replyType = ReplyType.THREAD
+                    UIKitConfig.groupChannelConfig.threadReplySelectType = ThreadReplySelectType.THREAD
+                    // set whether to use voice message
+                    UIKitConfig.groupChannelConfig.enableVoiceMessage = true
+                    // set typing indicator types
+                    UIKitConfig.groupChannelConfig.typingIndicatorTypes = setOf(TypingIndicatorType.BUBBLE, TypingIndicatorType.TEXT)
+                    // set whether to use feedback
+                    UIKitConfig.groupChannelConfig.enableFeedback = true
+                    // set custom params
+                    SendbirdUIKit.setCustomParamsHandler(object : CustomParamsHandler {
+                        override fun onBeforeCreateOpenChannel(params: OpenChannelCreateParams) {
+                            // You can set OpenChannelCreateParams globally before creating a open channel.
+                            params.customType = StringSet.SB_COMMUNITY_TYPE
+                        }
+                    })
+                }
+                SampleType.Notification -> {}
+                SampleType.Customization -> {}
+                SampleType.AiChatBot -> {
+                    // set typing indicator types
+                    UIKitConfig.groupChannelConfig.typingIndicatorTypes = setOf(TypingIndicatorType.BUBBLE)
+                    // set whether to use feedback
+                    UIKitConfig.groupChannelConfig.enableFeedback = true
+                }
+                else -> {
+                }
+            }
         }
     }
 
@@ -77,50 +129,5 @@ class BaseApplication : MultiDexApplication() {
 
         // setup uikit configurations
         setupConfigurations()
-    }
-
-    /**
-     * In a sample app, different contextual settings are used in a single app.
-     * These are only used in the sample, because if the app kills and resurrects due to low memory, the last used sample settings should be preserved.
-     */
-    private fun setupConfigurations() {
-        // set theme mode
-        SendbirdUIKit.setDefaultThemeMode(PreferenceUtils.themeMode)
-        // register push notification
-        SendbirdPushHelper.registerPushHandler(MyFirebaseMessagingService())
-        // set logger
-        SendbirdUIKit.setLogLevel(SendbirdUIKit.LogLevel.ALL)
-
-        when (PreferenceUtils.selectedSampleType) {
-            SampleType.Basic -> {
-                // set whether to use user profile
-                UIKitConfig.common.enableUsingDefaultUserProfile = true
-                // set whether to use typing indicators in channel list
-                UIKitConfig.groupChannelListConfig.enableTypingIndicator = true
-                // set whether to use read/delivery receipt in channel list
-                UIKitConfig.groupChannelListConfig.enableMessageReceiptStatus = true
-                // set whether to use user mention
-                UIKitConfig.groupChannelConfig.enableMention = true
-                // set reply type
-                UIKitConfig.groupChannelConfig.replyType = ReplyType.THREAD
-                UIKitConfig.groupChannelConfig.threadReplySelectType = ThreadReplySelectType.THREAD
-                // set whether to use voice message
-                UIKitConfig.groupChannelConfig.enableVoiceMessage = true
-                // set typing indicator types
-                UIKitConfig.groupChannelConfig.typingIndicatorTypes = setOf(TypingIndicatorType.BUBBLE, TypingIndicatorType.TEXT)
-
-                // set custom params
-                SendbirdUIKit.setCustomParamsHandler(object : CustomParamsHandler {
-                    override fun onBeforeCreateOpenChannel(params: OpenChannelCreateParams) {
-                        // You can set OpenChannelCreateParams globally before creating a open channel.
-                        params.customType = StringSet.SB_COMMUNITY_TYPE
-                    }
-                })
-            }
-            SampleType.Notification -> {}
-            SampleType.Customization -> {}
-            else -> {
-            }
-        }
     }
 }